// apps/admin/src/api/generated/model/feedbackCampaignConversationsDtoOutputCampaign.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Join The Six API
 * Operations API for Join The Six
 * OpenAPI spec version: 1.0.0
 */
import type { FeedbackCampaignConversationsDtoOutputCampaignStatus } from "./feedbackCampaignConversationsDtoOutputCampaignStatus";

export type FeedbackCampaignConversationsDtoOutputCampaign = {
  /**
   * @minimum 0
   * @maximum 9007199254740991
   */
  conversationCount: number;
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  eventId: string;
  /**
   * @minLength 1
   * @maxLength 200
   * @nullable
   */
  eventTitle: string | null;
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  id: string;
  /** @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$ */
  launchedAt: string;
  /**
   * @minimum 0
   * @maximum 9007199254740991
   */
  needsAttentionCount: number;
  /**
   * @minimum 0
   * @maximum 9007199254740991
   */
  openCount: number;
  /**
   * @maximum 9007199254740991
   * @exclusiveMinimum 0
   */
  questionSetVersion: number;
  status: FeedbackCampaignConversationsDtoOutputCampaignStatus;
};